TransferFunction,..Editor: move loading and saving tfs to editor
